+++ changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Eq` and `Ord` instances for `Operator`
+
 ### Deleted
 
 ## [0.3.0.0]